pyproject.toml
[project]
name = "adk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.23.0",
    "google-generativeai>=0.8.6",
    "python-dotenv>=1.2.1",
]
